+++ EPS HomeKit Bridge.indigoPlugin/Contents/Server Plugin/lib/plug.py
@@ -198,9 +198,10 @@ class plug:
 			
 			retval = self._callBack (AFTER, [valuesDict])
 			if retval is not None:
-				if "success" in retval: success = retval["success"]
-				if "valuesDict" in retval: valuesDict = retval["valuesDict"]
-				if "errorDict" in retval: errorDict = retval["errorDict"]
+				success, valuesDict, errorDict = retval
+				#if "success" in retval: success = retval["success"]
+				#if "valuesDict" in retval: valuesDict = retval["valuesDict"]
+				#if "errorDict" in retval: errorDict = retval["errorDict"]
 		
 		except Exception as e:
 			self.logger.error (ext.getException(e))